requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
gunicorn==20.1.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.34
Boto3
bcrypt
python-dotenv
awscli
google-auth
google-auth-oauthlib~=1.2.1
google-auth-httplib2
google-api-python-client~=2.146.0
botocore~=1.35.23
python-dateutil~=2.9.0.post0
protobuf~=5.28.1
flask_mail